restructure to split the implementation into multiple files
